feat(payment): PI-2467 Add "submit" method to the hosted-form-v2 package
