+++ packages/hosted-form-v2/src/hosted-form-factory.spec.ts
@@ -1,17 +1,12 @@
-import { PaymentIntegrationService } from '@bigcommerce/checkout-sdk/payment-integration-api';
-import { PaymentIntegrationServiceMock } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
-
 import HostedFieldType from './hosted-field-type';
 import HostedForm from './hosted-form';
 import HostedFormFactory from './hosted-form-factory';
 
 describe('HostedFormFactory', () => {
     let factory: HostedFormFactory;
-    let paymentIntegrationService: PaymentIntegrationService;
 
     beforeEach(() => {
-        paymentIntegrationService = new PaymentIntegrationServiceMock();
-        factory = new HostedFormFactory(paymentIntegrationService);
+        factory = new HostedFormFactory();
     });
 
     it('creates hosted form', () => {
